Collection

+++ lib/routes/homepage/widgets/collection_list.dart
@@ -6,22 +6,54 @@ import 'package:lingua_eidetic/models/collection.dart';
 import 'package:lingua_eidetic/routes/homepage/widgets/collection_card_v2.dart';
 import 'package:lingua_eidetic/services/collection_service.dart';
 
-class CollectionList extends StatelessWidget {
-  CollectionList({Key? key, required this.data}) : super(key: key);
+class CollectionList extends StatefulWidget {
+  CollectionList({
+    Key? key,
+    required this.data,
+    required this.query,
+  }) : super(key: key);
+  final String query;
   final Stream<QuerySnapshot<Object?>> data;
+
+  @override
+  _CollectionListState createState() => _CollectionListState();
+}
+
+class _CollectionListState extends State<CollectionList> {
   final CollectionService collectionService = CollectionService();
+  List<Collection> _cached = [];
+  List<String> _ids = [];
 
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<QuerySnapshot>(
-      stream: data,
+      stream: widget.data,
       builder: (context, snapshot) {
         if (snapshot.hasError) {
           return Text('Something went wrong');
         }
 
         if (snapshot.connectionState == ConnectionState.waiting) {
-          return Center(child: CircularProgressIndicator());
+          if (_cached.isEmpty)
+            return Center(child: CircularProgressIndicator());
+          return ListView.builder(
+            physics: NeverScrollableScrollPhysics(),
+            itemBuilder: (context, index) {
+              if (_cached[index].name.contains(widget.query))
+                return CollectionCardV2(
+                  key: Key(_ids[index]),
+                  title: _cached[index].name,
+                  avail: 5,
+                  total: 27,
+                  remove: () {
+                    collectionService.deleteCollection(
+                        collectionId: _ids[index]);
+                  },
+                );
+              return SizedBox();
+            },
+            itemCount: snapshot.data!.docs.length,
+          );
         }
 
         return ListView.builder(
@@ -31,15 +63,26 @@ class CollectionList extends StatelessWidget {
 
             Collection item = Collection.fromMap(
                 snapshot.data!.docs[index].data() as Map<String, dynamic>);
-            return CollectionCardV2(
-              key: Key(id),
-              title: item.name,
-              avail: 5,
-              total: 27,
-              remove: () {
-                collectionService.deleteCollection(collectionId: id);
-              },
-            );
+            if (index == 0) {
+              _cached.clear();
+              _ids.clear();
+            }
+            if (!_cached.contains(item)) {
+              _cached.add(item);
+              _ids.add(id);
+            }
+
+            if (item.name.contains(widget.query))
+              return CollectionCardV2(
+                key: Key(id),
+                title: item.name,
+                avail: 5,
+                total: 27,
+                remove: () {
+                  collectionService.deleteCollection(collectionId: id);
+                },
+              );
+            return SizedBox();
           },
           itemCount: snapshot.data!.docs.length,
         );
